Move the check_download_file routine into sunpy.util.net.
Also make it use the existing routines in that moudle.
